+++ _data/utils.js
@@ -1,4 +1,5 @@
-
+const fs = require('fs-extra');
+const path = require('path');
 
 function createNameForBattle(battle) {
   let label = battle.name;
@@ -26,6 +27,34 @@ function capitalize(s) {
   return s.charAt(0).toUpperCase() + s.slice(1)
 }
 
+async function useCache(main, cacheFilename) {
+  const isServing = process.env.ELEVENTY_SERVE === 'true';
+  const cacheFilePath = path.resolve(__dirname, '../_cache/' + cacheFilename);
+  let dataInCache = null;
+
+  if (isServing && await fs.pathExists(cacheFilePath)) {
+    // Read file from cache.
+    dataInCache = await fs.readJSON(cacheFilePath);
+    console.log('Using from cache: ' + cacheFilename);
+  }
+
+  if (!dataInCache) {
+    const result = await main();
+
+    if (isServing) {
+      // Write data to cache.
+      fs.writeJSON(cacheFilePath, result, err => {
+        if (err) {console.error(err)}
+      });
+    }
+
+    dataInCache = result;
+  }
+
+  return dataInCache;
+}
+
 module.exports = {
-  createNameForBattle
+  createNameForBattle,
+  useCache
 };